refactor error type enum as a common entry set enum
